Update Order for work hour tests
