--- UI.NUnitVersion/Admin/WorkAllocation/UploadWorkHourTests.cs
@@ -1,31 +1,0 @@
-namespace UI.NUnitVersion.Admin.WorkAllocation;
-
-[Category("Daily")]
-public class UploadWorkHourTests : AdminWebUiTest
-{
-    [Test]
-    public void UploadValidWorkHours()
-    {
-        var driver = VhDriver.GetDriver();
-        driver.Navigate().GoToUrl(EnvConfigSettings.AdminUrl);
-        var loginPage = new AdminWebLoginPage(driver, EnvConfigSettings.DefaultElementWait);
-        var dashboardPage = loginPage.Login(AdminLoginUsername, EnvConfigSettings.UserPassword);
-
-        var workAllocationPage = dashboardPage.GoToManageWorkAllocation();
-        workAllocationPage.UploadWorkHoursFile(Path.Join("TestData", "GoodWorkHours.csv"));
-        Assert.Pass();
-    }
-
-    [Test]
-    public void UploadValidNonAvailabilityHours()
-    {
-        var driver = VhDriver.GetDriver();
-        driver.Navigate().GoToUrl(EnvConfigSettings.AdminUrl);
-        var loginPage = new AdminWebLoginPage(driver, EnvConfigSettings.DefaultElementWait);
-        var dashboardPage = loginPage.Login(AdminLoginUsername, EnvConfigSettings.UserPassword);
-
-        var workAllocationPage = dashboardPage.GoToManageWorkAllocation();
-        workAllocationPage.UploadNonWorkHoursFile(Path.Join("TestData", "GoodNonAvailabilityHours.csv"));
-        Assert.Pass();
-    }
-}
